+++ src/notebooks/export/exportInterpreterFinder.node.ts
@@ -4,8 +4,8 @@ import {
     IJupyterInterpreterDependencyManager,
     INbConvertInterpreterDependencyChecker
 } from '../../kernels/jupyter/types';
-import * as localize from '../common/utils/localize';
-import { PythonEnvironment } from '../pythonEnvironments/info';
+import * as localize from '../../platform/common/utils/localize';
+import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
 
 @injectable()
 export class ExportInterpreterFinder {
