+++ src/notebooks/export/exportToPDF.ts
@@ -1,17 +1,18 @@
 import { inject, injectable } from 'inversify';
 import { CancellationToken, NotebookDocument, Uri } from 'vscode';
-import { PythonEnvironment } from '../pythonEnvironments/info';
+import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
 import { ExportFormat, IExportBase, INbConvertExport } from './types';
 
 @injectable()
-export class ExportToHTML implements INbConvertExport {
+export class ExportToPDF implements INbConvertExport {
     constructor(@inject(IExportBase) protected readonly exportBase: IExportBase) {}
+
     public async export(
         sourceDocument: NotebookDocument,
         target: Uri,
         interpreter: PythonEnvironment,
         token: CancellationToken
     ): Promise<void> {
-        await this.exportBase.executeCommand(sourceDocument, target, ExportFormat.html, interpreter, token);
+        await this.exportBase.executeCommand(sourceDocument, target, ExportFormat.pdf, interpreter, token);
     }
 }
