+++ src/notebooks/export/fileConverter.ts
@@ -6,14 +6,14 @@
 import { inject, injectable, named } from 'inversify';
 import { CancellationToken, NotebookDocument, Uri, workspace } from 'vscode';
 import { sendTelemetryEvent } from '../../telemetry';
-import { IApplicationShell } from '../common/application/types';
-import { Telemetry } from '../common/constants';
-import { IConfigurationService } from '../common/types';
-import * as localize from '../common/utils/localize';
-import { noop } from '../common/utils/misc';
-import { traceError } from '../logging';
-import { ProgressReporter } from '../progress/progressReporter';
-import { PythonEnvironment } from '../pythonEnvironments/info';
+import { IApplicationShell } from '../../platform/common/application/types';
+import { Telemetry } from '../../platform/common/constants';
+import { IConfigurationService } from '../../platform/common/types';
+import * as localize from '../../platform/common/utils/localize';
+import { noop } from '../../platform/common/utils/misc';
+import { traceError } from '../../platform/logging';
+import { ProgressReporter } from '../../platform/progress/progressReporter';
+import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
 import { ExportFileOpener } from './exportFileOpener';
 import { ExportUtilBase } from './exportUtil';
 import { ExportFormat, IExport, IExportDialog, IFileConverter, INbConvertExport } from './types';
