+++ src/platform/serviceRegistry.node.ts
@@ -12,17 +12,6 @@ import { DataScienceStartupTime } from './common/constants';
 import { IExtensionSingleActivationService, IExtensionSyncActivationService } from './activation/types';
 import { PreReleaseChecker } from './common/prereleaseChecker.node';
 import { IConfigurationService, IDataScienceCommandListener, IExtensionContext } from './common/types';
-import { ExportBase } from './export/exportBase.node';
-import { ExportDialog } from './export/exportDialog';
-import { ExportFileOpener } from './export/exportFileOpener';
-import { ExportInterpreterFinder } from './export/exportInterpreterFinder.node';
-import { ExportToHTML } from './export/exportToHTML';
-import { ExportToPDF } from './export/exportToPDF';
-import { ExportToPython } from './export/exportToPython';
-import { ExportToPythonPlain } from './export/exportToPythonPlain';
-import { ExportUtil } from './export/exportUtil.node';
-import { FileConverter } from './export/fileConverter.node';
-import { IFileConverter, INbConvertExport, ExportFormat, IExport, IExportDialog, IExportBase } from './export/types';
 import { KernelProgressReporter } from './progress/kernelProgressReporter';
 import { ProgressReporter } from './progress/progressReporter';
 import { StatusProvider } from './progress/statusProvider';
@@ -36,7 +25,6 @@ import { IFileSystemNode } from './common/platform/types.node';
 import { FileSystem } from './common/platform/fileSystem.node';
 import { WorkspaceService } from './common/application/workspace.node';
 import { OutputCommandListener } from './logging/outputCommandListener';
-import { ExportUtilBase } from './export/exportUtil';
 
 export function registerTypes(context: IExtensionContext, serviceManager: IServiceManager, isDevMode: boolean) {
     serviceManager.addSingleton<FileSystem>(FileSystem, FileSystem);
@@ -58,19 +46,6 @@ export function registerTypes(context: IExtensionContext, serviceManager: IServi
 
     serviceManager.addSingleton<IStatusProvider>(IStatusProvider, StatusProvider);
     serviceManager.addSingleton<ProgressReporter>(ProgressReporter, ProgressReporter);
-    serviceManager.addSingleton<IFileConverter>(IFileConverter, FileConverter);
-    serviceManager.addSingleton<ExportInterpreterFinder>(ExportInterpreterFinder, ExportInterpreterFinder);
-    serviceManager.addSingleton<ExportFileOpener>(ExportFileOpener, ExportFileOpener);
-
-    serviceManager.addSingleton<IExportBase>(IExportBase, ExportBase);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToPDF, ExportFormat.pdf);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToHTML, ExportFormat.html);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToPython, ExportFormat.python);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportBase, 'Export Base');
-    serviceManager.addSingleton<IExport>(IExport, ExportToPythonPlain, ExportFormat.python);
-    serviceManager.addSingleton<ExportUtilBase>(ExportUtilBase, ExportUtilBase);
-    serviceManager.addSingleton<ExportUtil>(ExportUtil, ExportUtil);
-    serviceManager.addSingleton<IExportDialog>(IExportDialog, ExportDialog);
     serviceManager.addSingleton<IExtensionSyncActivationService>(
         IExtensionSyncActivationService,
         KernelProgressReporter
