+++ src/platform/serviceRegistry.web.ts
@@ -24,18 +24,9 @@ import { IStatusProvider } from './progress/types';
 import { WorkspaceService } from './common/application/workspace.web';
 import { IExtensionSyncActivationService } from './activation/types';
 import { OutputCommandListener } from './logging/outputCommandListener';
-import { ExportDialog } from './export/exportDialog';
-import { ExportFormat, IExport, IExportBase, IExportDialog, IFileConverter, INbConvertExport } from './export/types';
-import { FileConverter } from './export/fileConverter';
-import { ExportFileOpener } from './export/exportFileOpener';
-import { ExportToPythonPlain } from './export/exportToPythonPlain';
+
 import { IFileSystem } from './common/platform/types';
 import { FileSystem } from './common/platform/fileSystem';
-import { ExportBase } from './export/exportBase.web';
-import { ExportUtilBase } from './export/exportUtil';
-import { ExportToHTML } from './export/exportToHTML';
-import { ExportToPDF } from './export/exportToPDF';
-import { ExportToPython } from './export/exportToPython';
 import { KernelProgressReporter } from './progress/kernelProgressReporter';
 
 export function registerTypes(context: IExtensionContext, serviceManager: IServiceManager, isDevMode: boolean) {
@@ -47,16 +38,7 @@ export function registerTypes(context: IExtensionContext, serviceManager: IServi
     serviceManager.addSingleton<IConfigurationService>(IConfigurationService, ConfigurationService);
     serviceManager.addSingleton<IStatusProvider>(IStatusProvider, StatusProvider);
     serviceManager.addSingleton<IDataScienceCommandListener>(IDataScienceCommandListener, OutputCommandListener);
-    serviceManager.addSingleton<ExportFileOpener>(ExportFileOpener, ExportFileOpener);
-    serviceManager.addSingleton<IExportBase>(IExportBase, ExportBase);
-    serviceManager.addSingleton<IExportDialog>(IExportDialog, ExportDialog);
     serviceManager.addSingleton<ProgressReporter>(ProgressReporter, ProgressReporter);
-    serviceManager.addSingleton<IFileConverter>(IFileConverter, FileConverter);
-    serviceManager.addSingleton<IExport>(IExport, ExportToPythonPlain, ExportFormat.python);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToHTML, ExportFormat.html);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToPDF, ExportFormat.pdf);
-    serviceManager.addSingleton<INbConvertExport>(INbConvertExport, ExportToPython, ExportFormat.python);
-    serviceManager.addSingleton<ExportUtilBase>(ExportUtilBase, ExportUtilBase);
 
     registerCommonTypes(serviceManager);
     registerApiTypes(serviceManager);
