+++ src/telemetry/types.ts
@@ -16,7 +16,7 @@ import { CheckboxState, EventName, PlatformErrors, SliceOperationSource } from '
 import { DebuggingTelemetry } from '../kernels/debugger/constants';
 import { EnvironmentType, PythonEnvironment } from '../platform/pythonEnvironments/info';
 import { TelemetryErrorProperties, ErrorCategory } from '../platform/errors/types';
-import { ExportFormat } from '../platform/export/types';
+import { ExportFormat } from '../notebooks/export/types';
 import { InterruptResult, KernelConnectionMetadata, KernelInterpreterDependencyResponse } from '../kernels/types';
 import { IExportedKernelService } from '../platform/api/extension';
 import { PreferredKernelExactMatchReason } from '../notebooks/controllers/notebookControllerManager';
