+++ src/test/datascience/export/exportToPython.vscode.test.ts
@@ -7,8 +7,8 @@ import * as path from '../../../platform/vscode-path/path';
 import { CancellationTokenSource, Uri, workspace } from 'vscode';
 import { IDocumentManager } from '../../../platform/common/application/types';
 import { IFileSystemNode } from '../../../platform/common/platform/types.node';
-import { ExportInterpreterFinder } from '../../../platform/export/exportInterpreterFinder.node';
-import { INbConvertExport, ExportFormat } from '../../../platform/export/types';
+import { ExportInterpreterFinder } from '../../../notebooks/export/exportInterpreterFinder.node';
+import { INbConvertExport, ExportFormat } from '../../../notebooks/export/types';
 import { IExtensionTestApi } from '../../common.node';
 import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../constants.node';
 import { closeActiveWindows, initialize } from '../../initialize.node';
